requirements.txt
# python version 3.10
drake==1.23.0
scipy==1.11.4
tqdm==4.66.1
matplotlib==3.8.2